#204: Make Availabilities component dynamic
