+++ Frontend/ShoppingList.Frontend.Models/Items/Models/ItemType.cs
@@ -1,11 +1,10 @@
-﻿using ProjectHermes.ShoppingList.Frontend.Models.Stores.Models;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 
 namespace ProjectHermes.ShoppingList.Frontend.Models.Items.Models
 {
-    public class ItemType
+    public class ItemType : IAvailable
     {
         public ItemType(Guid id, string name, IEnumerable<ItemAvailability> availabilities)
         {
@@ -17,17 +16,5 @@ namespace ProjectHermes.ShoppingList.Frontend.Models.Items.Models
         public Guid Id { get; }
         public string Name { get; set; }
         public List<ItemAvailability> Availabilities { get; }
-
-        public IEnumerable<ItemStore> GetNotRegisteredStores(IEnumerable<Store> stores)
-        {
-            var registeredStoreIds = Availabilities.Select(av => av.Store.Id).OrderBy(id => id);
-            var allStoreIds = stores.Select(s => s.Id).OrderBy(id => id).ToList();
-
-            if (allStoreIds.SequenceEqual(registeredStoreIds))
-                return Enumerable.Empty<ItemStore>();
-
-            var availableStoreIds = allStoreIds.Except(registeredStoreIds).ToList();
-            return stores.Where(s => availableStoreIds.Contains(s.Id)).Select(s => s.AsStoreItemStore());
-        }
     }
 }
