+++ Frontend/ShoppingList.Frontend.Models/Stores/Models/Store.cs
@@ -84,7 +84,7 @@ namespace ProjectHermes.ShoppingList.Frontend.Models.Stores.Models
             }
         }
 
-        public ItemStore AsStoreItemStore()
+        public ItemStore AsItemStore()
         {
             return new ItemStore(Id, Name, Sections.Select(s => s.AsStoreItemSection()));
         }
