+++ Frontend/ShoppingList.Frontend.WebApp/Components/Index/TemporaryItemCreator.razor
@@ -7,7 +7,7 @@
 @inject ITemporaryItemCreationService temporaryItemCreationService;
 @inject IShoppingListApiService apiService;
 
-<div>
+<div class="temporary-item-creator">
     <Button Type="@ButtonType.Primary"
             OnClick="@OnOpenCreationButtonClicked"
             Class="d-flex align-items-center"
