+++ Frontend/ShoppingList.Frontend.WebApp/Components/Items/Editor/Availabilities.razor
@@ -5,16 +5,18 @@
     <div class="mt-3 d-flex justify-content-start align-items-center">
         <div class="font-weight-bold mr-2">Stores</div>
         <Button class="d-flex align-items-center" Size="@AntSizeLDSType.Small" OnClick="OnAddStoreButtonClicked"
-                Disabled="@(!State.EditedItem.GetNotRegisteredStores(State.Stores).Any())">
+                Disabled="@(!Available.GetNotRegisteredStores(State.Stores).Any())">
             <Icon Type="plus"></Icon>
         </Button>
     </div>
     <div class="py-2">
-        @for (int i = 0; i < State.EditedItem.Availabilities.Count; i++)
+        @for (int i = 0; i < Available.Availabilities.Count; i++)
         {
             <div class="pb-3">
                 <Availability State="@State"
-                              ComponentIndex="i"></Availability>
+                              Model="@Available.Availabilities[i]"
+                              Index="i"
+                              OnRemove="index => Available.Availabilities.RemoveAt(index)"></Availability>
             </div>
         }
     </div>
@@ -22,16 +24,16 @@
 
 @code {
     [Parameter] public ItemsState State { get; set; }
+    [Parameter] public IAvailable Available { get; set; }
 
     private void OnAddStoreButtonClicked()
     {
-        var availableStores = State.EditedItem.GetNotRegisteredStores(State.Stores);
+        var availableStores = Available.GetNotRegisteredStores(State.Stores);
         if (!availableStores.Any())
             return;
-
+        
         var chosenStore = availableStores.First();
-        State.EditedItem.Availabilities.Add(
-            new ItemAvailability(chosenStore, 1, GetSectionForStoreId(chosenStore.Id).Id));
+        Available.Availabilities.Add(new ItemAvailability(chosenStore, 1, GetSectionForStoreId(chosenStore.Id).Id));
 
         State.StateChanged();
     }
