--- Frontend/ShoppingList.Frontend.WebApp/Components/Items/Editor/ItemTypeAvailabilities.razor
@@ -1,48 +1,0 @@
-﻿@using ProjectHermes.ShoppingList.Frontend.Models.Items;
-@using ProjectHermes.ShoppingList.Frontend.Models.Items.Models
-
-<div>
-    <div class="mt-3 d-flex justify-content-start align-items-center">
-        <div class="mr-2">Stores</div>
-        <Button class="d-flex align-items-center" Size="@AntSizeLDSType.Small" OnClick="OnAddStoreButtonClicked"
-                Disabled="@(!State.EditedItem.ItemTypes[ItemTypeIndex].GetNotRegisteredStores(State.Stores).Any())">
-            <Icon Type="plus"></Icon>
-        </Button>
-    </div>
-    <div class="d-flex flex-row py-2">
-        <div>
-            @for (int i = 0; i < State.EditedItem.ItemTypes[ItemTypeIndex].Availabilities.Count; i++)
-            {
-                <div class="pb-1">
-                    <ItemTypeAvailability State="@State"
-                                          ItemTypeIndex="ItemTypeIndex"
-                                          ComponentIndex="i"></ItemTypeAvailability>
-                </div>
-            }
-        </div>
-    </div>
-</div>
-
-@code {
-    [Parameter] public ItemsState State { get; set; }
-    [Parameter] public int ItemTypeIndex { get; set; }
-
-    private void OnAddStoreButtonClicked()
-    {
-        var availableStores = State.EditedItem.ItemTypes[ItemTypeIndex].GetNotRegisteredStores(State.Stores);
-        if (!availableStores.Any())
-            return;
-
-        var chosenStore = availableStores.First();
-        State.EditedItem.ItemTypes[ItemTypeIndex].Availabilities.Add(
-            new ItemAvailability(chosenStore, 1, GetSectionForStoreId(chosenStore.Id).Id));
-
-        StateHasChanged();
-    }
-
-    private ItemSection GetSectionForStoreId(Guid storeId)
-    {
-        var section = State.Stores.Single(s => s.Id == storeId).Sections.Single(s => s.IsDefaultSection);
-        return section.AsStoreItemSection();
-    }
-}
