--- Frontend/ShoppingList.Frontend.WebApp/Components/Items/Editor/ItemTypeAvailability.razor
@@ -1,68 +1,0 @@
-﻿@using ProjectHermes.ShoppingList.Frontend.Models.Items;
-@using ProjectHermes.ShoppingList.Frontend.Models;
-@using ProjectHermes.ShoppingList.Frontend.Models.Items.Models
-@using ProjectHermes.ShoppingList.Frontend.Models.Stores.Models
-
-<div class="d-flex flex-row align-items-center">
-    <div class="max-w-80">
-        <Input @bind-Value="@availability.PricePerQuantity" />
-    </div>
-    <div class="mr-2">@(State.EditedItem.QuantityType.PriceLabel)</div>
-    <Select Class="mr-2"
-            DataSource="State.Stores.ToList()"
-            TItem="Store"
-            TItemValue="Guid"
-            DefaultValue="availability.Store.Id"
-            ValueName="@nameof(Store.Id)"
-            LabelName="@nameof(Store.Name)"
-            OnSelectedItemChanged="@OnStoreChanged">
-    </Select>
-    <SingleSelect TItem="ItemSection"
-                  TValue="Guid"
-                  ValueName="@nameof(ItemSection.Id)"
-                  LabelName="@nameof(ItemSection.Name)"
-                  DataSource="@availability.Store.Sections.ToList()"
-                  OnItemChanged="OnDefaultSectionChanged"
-                  DefaultValue="@availability.DefaultSectionId"></SingleSelect>
-    <div class="ml-2">
-        <Button Class="d-flex align-items-center"
-                OnClick="@OnRemoveStoreButtonClicked"
-                Danger="true">
-            <Icon Type="minus"></Icon>
-        </Button>
-    </div>
-</div>
-
-@code {
-    [Parameter] public ItemsState State { get; set; }
-    [Parameter] public int ItemTypeIndex { get; set; }
-    [Parameter] public int ComponentIndex { get; set; }
-
-    private ItemAvailability availability;
-
-    protected override void OnInitialized()
-    {
-        base.OnInitialized();
-        availability = State.EditedItem.ItemTypes[ItemTypeIndex].Availabilities[ComponentIndex];
-    }
-
-    private void OnStoreChanged(Store store)
-    {
-        availability.Store = store.AsStoreItemStore();
-
-        var defaultSection = store.Sections.Single(s => s.IsDefaultSection);
-        availability.ChangeDefaultSectionId(defaultSection.Id.BackendId);
-        StateHasChanged();
-    }
-
-    private void OnRemoveStoreButtonClicked()
-    {
-        State.EditedItem.ItemTypes[ItemTypeIndex].Availabilities.RemoveAt(ComponentIndex);
-        State.StateChanged();
-    }
-
-    private void OnDefaultSectionChanged(ItemSection section)
-    {
-        availability.ChangeDefaultSectionId(section.Id);
-    }
-}
