+++ Frontend/ShoppingList.Frontend.WebApp/wwwroot/css/components/index/temporary-item-creator.css
@@ -1,3 +1,3 @@
-﻿.price-input {
+﻿.temporary-item-creator .price-input {
     max-width: 65px;
 }
